:sparkles: cruiser naming 변경
