+++ src/main/java/kr/co/antoon/common/domain/cruiser/Cruiser.java
@@ -1,0 +1,10 @@
+package kr.co.antoon.common.domain.cruiser;
+
+public interface Cruiser {
+    /**
+     * Cruiser For Message
+     *
+     * @param content 알림 메세지
+     **/
+    void send(String content);
+}
