+++ src/main/java/kr/co/antoon/common/dto/SlackCruiserDto.java
@@ -1,0 +1,3 @@
+package kr.co.antoon.common.dto;
+
+public record SlackCruiserDto(String text) { }
