--- src/main/java/kr/co/antoon/common/dto/SlackCrusierDto.java
@@ -1,3 +1,0 @@
-package kr.co.antoon.common.dto;
-
-public record SlackCrusierDto(String text) { }
